requirements.txt
openai
python-dotenv
pydub
vosk
requests
watchdog
